+++ src/main/java/com/github/wintersteve25/tau/theme/Theme.java
@@ -4,6 +4,7 @@ import com.github.wintersteve25.tau.utils.Color;
 import com.github.wintersteve25.tau.utils.InteractableState;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.network.chat.FormattedText;
 
 import java.util.List;
@@ -15,7 +16,7 @@ public interface Theme {
     
     void drawScrollbar(PoseStack poseStack, int x, int y, int width, int height, float partialTicks, int mouseX, int mouseY);
    
-    void drawTooltip(PoseStack poseStack, int mouseX, int mouseY, int screenWidth, int screenHeight, Font fontRenderer, List<FormattedText> tooltips);
+    void drawTooltip(PoseStack poseStack, int mouseX, int mouseY, int screenWidth, int screenHeight, Font fontRenderer, List<ClientTooltipComponent> tooltips);
     
     Color getTextColor();
 }